Revert "V2 Auto Multiline Detection Foundation - Aggregator, Labeler, and Ref…"
This reverts commit 68973179ebb2b4c3fc491db184a73b1a0579b19f.
